refactor: 인증 쿠키 캡슐화
