+++ backend/src/main/java/site/coduo/member/controller/dto/auth/SignInWebResponse.java
@@ -1,4 +1,4 @@
-package site.coduo.member.controller.dto.member;
+package site.coduo.member.controller.dto.auth;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import site.coduo.member.service.dto.SignInServiceResponse;
