+++ backend/src/main/java/site/coduo/member/controller/dto/auth/SignUpRequest.java
@@ -1,4 +1,4 @@
-package site.coduo.member.controller.dto.member;
+package site.coduo.member.controller.dto.auth;
 
 public record SignUpRequest(String username) {
 }
